% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbat4r-package.R
\docType{package}
\name{cbat4r-package}
\alias{cbat4r}
\alias{cbat4r-package}
\title{cbat4r: Scaffold jsPsych Tasks in the CBAT Format}
\description{
Creates the directory layout and starter files to run 'jsPsych' experiments in the CBAT (Cognitive/Behavioral Assessment Tasks) format: HTML entry points for the Demo, JATOS and CEMA environments, init/run scripts, starter task code, and the official 'jsPsych' distribution (downloaded once per version and cached). Also generates Likert questionnaire and informed-consent tasks, Phaser 3 starters, and packages studies into JATOS '.jzip' archives.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cba-toolbox/cbat4r}
  \item Report bugs at \url{https://github.com/cba-toolbox/cbat4r/issues}
}

}
\author{
\strong{Maintainer}: Yoshihiko Kunisato \email{ykunisato@psy.senshu-u.ac.jp}

Authors:
\itemize{
  \item Yoshihiko Kunisato \email{ykunisato@psy.senshu-u.ac.jp}
}

}
\keyword{internal}
