% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cbat.R
\name{set_ic}
\alias{set_ic}
\title{Set jsPsych Informed Consent Task}
\usage{
set_ic(
  task_name = "ic",
  ic_markdown,
  ic_question =
    "Do you agree to participate in the research after reading and understanding the above information?",
  ic_agree_label =
    "I have read and understood the information and agree to participate in the research.",
  jsPsych_version = "8.2.2"
)
}
\arguments{
\item{task_name}{A character string specifying the name of the task.}

\item{ic_markdown}{A character string containing the IC text in Markdown format, or a path to a .md file.}

\item{ic_question}{A character string for the consent question. Default is English.}

\item{ic_agree_label}{A character string for the consent checkbox label. Default is English.}

\item{jsPsych_version}{A character string specifying the version of jsPsych to use (e.g., "7.3.4" or "8.2.2").}
}
\description{
This function creates a directory and prepares necessary files to run an Informed Consent task using jsPsych. It converts Markdown text to HTML for the consent document.
}
