% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cbat.R
\name{set_qnr}
\alias{set_qnr}
\title{Set jsPsych Questionnaire (Likert) Task}
\usage{
set_qnr(
  task_name = "task_name",
  scale,
  item,
  instruction,
  randomize_order = "false",
  jsPsych_version = "8.2.2"
)
}
\arguments{
\item{task_name}{A character string specifying the name of the task.}

\item{scale}{A character vector defining the default scale labels.}

\item{item}{A data frame defining the questionnaire items.}

\item{instruction}{A character string specifying the instruction text.}

\item{randomize_order}{A character string ("true" or "false"). Default is "false".}

\item{jsPsych_version}{A character string specifying the version of jsPsych to use.}
}
\description{
This function creates a directory and prepares necessary files (HTML, JS, CSS) to run a questionnaire task using jsPsych.
}
