% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cbat.R
\name{set_qnr}
\alias{set_qnr}
\title{Set jsPsych Questionnaire (Likert) Task}
\usage{
set_qnr(
  task_name = "task_name",
  scale,
  item,
  instruction,
  randomize_order = "false",
  jsPsych_version = "8.2.2"
)
}
\arguments{
\item{task_name}{A character string specifying the name of the task. This will be used for the directory and file names. Default is "task_name".}

\item{scale}{A character vector defining the default scale labels (e.g., c("Strongly Disagree", "Disagree", ...)).}

\item{item}{A data frame defining the questionnaire items. It must contain columns: 'prompt' (question text), 'required' (true/false), 'name' (variable name), and 'labels' (variable name for scale labels).}

\item{instruction}{A character string specifying the instruction or preamble text to be displayed at the top of the questionnaire.}

\item{randomize_order}{A character string ("true" or "false") indicating whether to randomize the order of questions. Default is "false".}

\item{jsPsych_version}{A character string specifying the version of jsPsych to use (e.g., "7.3.4" or "8.2.2").
The function checks if the version is supported and downloads the corresponding release.}
}
\description{
This function creates a directory and prepares necessary files (HTML, JS, CSS) to run a questionnaire task using jsPsych. It supports specific versions of jsPsych (v7 and v8).
}
\examples{
\dontrun{
  scale_list <- c("Strongly Disagree", "Disagree", "Neutral", "Agree", "Strongly Agree")
  items <- data.frame(
    prompt = c("I feel happy.", "I feel energetic."),
    required = c("true", "true"),
    name = c("happy", "energetic"),
    labels = c("scale", "scale")
  )
  set_qnr(task_name = "mood_survey",
          scale = scale_list,
          item = items,
          instruction = "Please answer the following questions.",
          jsPsych_version = "8.2.2")
}
}
